requirements.txt
Flask==2.2.2
requests==2.28.1
scikit-learn==1.1.2
python-dotenv==0.20.0
openai==0.15.0
